requirements.txt
openai==1.55.0
python-dotenv==1.0.1
pydantic-settings==2.6.1
pydantic==2.10.1
pathspec==0.12.1
httpx==0.27.2